Hotfix0.2.0

+++ services/candidate.go
@@ -155,7 +155,7 @@ func (s CandidateService) DelegatorCandidateList(reqVO vo.DelegatorCandidateList
 func (s CandidateService) buildCandidates(
 	cd document.Candidate,
 	delegator []document.Delegator,
-	totalShares uint64,
+	totalShares float64,
 ) document.Candidate {
 
 	delegators := make([]document.Delegator, 0)
@@ -167,7 +167,7 @@ func (s CandidateService) buildCandidates(
 		}
 	}
 	if totalShares != 0 {
-		cd.VotingPower = float64(cd.Shares) / float64(totalShares)
+		cd.VotingPower = float64(cd.Shares) / totalShares
 	}
 
 
@@ -175,6 +175,6 @@ func (s CandidateService) buildCandidates(
 }
 
 // get total shares
-func (s CandidateService) getTotalShares() (uint64, error) {
+func (s CandidateService) getTotalShares() (float64, error) {
 	return candidateModel.GetTotalShares()
 }
